\name{lambdaGenPick}
\alias{lambdaGenPick}
\title{Auxiliary function to compute Segers' estimator}
\description{
This function computes

\deqn{\lambda_{\delta, \rho}^c}

given in Theorem 4.1 of Segers (2005) and is called by \code{\link{generalizedPick}}.
It is not intended to be called by the user.
}
\references{
Mueller, S. and Rufibach K. (2009).
Smooth tail index estimation.
\emph{J. Stat. Comput. Simul.}, \bold{79}, 1155--1167.

Segers, J. (2005).
Generalized Pickands estimators for the extreme value index.
\emph{J. Statist. Plann. Inference}, \bold{128}, 381--396.
}
\author{
Kaspar Rufibach (maintainer), \email{kaspar.rufibach@gmail.com}, \cr \url{http://www.kasparrufibach.ch} 

Samuel Mueller, \email{samuel.mueller@sydney.edu.au}, \cr \url{www.maths.usyd.edu.au/ut/people?who=S_Mueller}

Kaspar Rufibach acknowledges support by the Swiss National Science Foundation SNF, \url{http://www.snf.ch}
}
\seealso{Called by \code{\link{generalizedPick}}.}

\keyword{distribution}
\keyword{htest}
\keyword{nonparametric}


